requirements.txt
pillow
djangorestframework